[metadata] Move _MonoClass definition to class-private-definition.h
