+++ msvc/libmonoruntime-common.targets
@@ -18,6 +18,7 @@
     <ClCompile Include="$(MonoSourceLocation)\mono\metadata\class-init.c" />
     <ClInclude Include="$(MonoSourceLocation)\mono\metadata\class-init.h" />
     <ClCompile Include="$(MonoSourceLocation)\mono\metadata\class-accessors.c" />
+    <ClInclude Include="$(MonosourceLocation)\mono\metadata\class-private-definition.h" />
     <ClCompile Include="$(MonoSourceLocation)\mono\metadata\cominterop.c" />
     <ClInclude Include="$(MonoSourceLocation)\mono\metadata\cominterop.h" />
     <ClInclude Include="$(MonoSourceLocation)\mono\metadata\console-io.h" />
